Fix tests, issue with quotes, update deps
This commit does three things:
1) Fix the test case
2) Fixes an issue with quotes in CSS strings
3) Updates dependencies to fix NPM audit warnings

1 - Fix the test case
The test case was calling assert() instead of assert.equals(). assert()
always passes if the first arg is truthy.

I also moved the expected code out to a JS file and ran that through the
same rollup pipeline.  I feel like this might make the test a bit less
brittle.

2 - Issue with quotes in CSS strings
If there is a quote (") in the CSS string, like in the test case, the
regex would use that as the start of the CSS string, not the actual
start. Before this fix, the fixed test case didn't pass.

3 - Update dependencies
There were quite a few audit warnings when I installed, so I ran npm update, which fixed them.